More Structure added to atom
The recursive nature of the data structure is represented by the enum
AtomNodes. At any level, a list of atoms is represented as a
Vec<AtomNodes>

Container or leaf atoms are further separated into their own enums and
each atom is represented by it own struct.

All the types support std::fmt::Display which is what is used to
construct the display strings.